requirements.txt
numpy>=1.24
opencv-python>=4.8
matplotlib>=3.7
pillow>=10.0
pytesseract>=0.3.10
scikit-learn>=1.3
torch>=2.2; platform_system != "Darwin" or platform_machine != "arm64"
torchvision>=0.17; platform_system != "Darwin" or platform_machine != "arm64"

